+++ mgm/XrdMgmOfs/fsctl/Txstate.cc
@@ -45,50 +45,46 @@
   if (txid)
   {
     long long id = strtoll(txid, 0, 10);
-    if (sprogress)
-    {
-      if (sprogress)
-      {
+
+    if (sprogress) {
+      if (sprogress) {
         float progress = atof(sprogress);
-        if (!gTransferEngine.SetProgress(id, progress))
-        {
-          eos_thread_err("unable to set progress for transfer id=%lld progress=%.02f", id, progress);
-          return Emsg(epname, error, ENOENT, "set transfer state - transfer has been canceled [EIDRM]", "");
-        }
-        else
-        {
+
+        if (!gTransferEngine.SetProgress(id, progress)) {
+          eos_thread_err("unable to set progress for transfer id=%lld progress=%.02f", id,
+          progress);
+          return Emsg(epname, error, ENOENT,
+          "set transfer state - transfer has been canceled [EIDRM]", "");
+        } else {
           eos_thread_info("id=%lld progress=%.02f", id, progress);
         }
       }
     }
 
-    if (sstate)
-    {
+    if (sstate) {
       char* logout = 0;
-      unsigned loglen = 0;
-      if (logb64)
-      {
+      size_t loglen = 0;
+
+      if (logb64) {
         XrdOucString slogb64 = logb64;
 
-        if (eos::common::SymKey::Base64Decode(slogb64, logout, loglen))
-        {
+        if (eos::common::SymKey::Base64Decode(slogb64, logout, loglen)) {
           logout[loglen] = 0;
-          if (!gTransferEngine.SetLog(id, logout))
-          {
+
+          if (!gTransferEngine.SetLog(id, logout)) {
             eos_thread_err("unable to set log for transfer id=%lld", id);
           }
         }
       }
 
       int state = atoi(sstate);
-      if (!gTransferEngine.SetState(id, state))
-      {
+
+      if (!gTransferEngine.SetState(id, state)) {
         eos_thread_err("unable to set state for transfer id=%lld state=%s",
                        id, TransferEngine::GetTransferState(state));
-      }
-      else
-      {
-        eos_thread_info("id=%lld state=%s", id, TransferEngine::GetTransferState(state));
+      } else {
+        eos_thread_info("id=%lld state=%s", id,
+                        TransferEngine::GetTransferState(state));
       }
     }
   }
